@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -25,6 +25,7 @@ dependencies:
   google_sign_in: ^4.0.7
   cloud_firestore: ^0.12.9+5
   intl: ^0.16.0
+  flutter_svg: 0.15.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -50,8 +51,8 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #  - images/a_dot_burr.jpeg
+  assets:
+    - assets/images/
   #  - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
@@ -79,3 +80,17 @@ flutter:
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
+
+  fonts:
+    - family: afek
+      fonts:
+        - asset: assets/fonts/afek-bold-aaa.otf
+          weight: 700
+        - asset: assets/fonts/afek-medium-aaa.otf
+        - asset: assets/fonts/afek-regular-aaa.otf
+    - family: sfProDisplay
+      fonts:
+        - asset: assets/fonts/SFProDisplay-Regular.ttf
+    - family: Heebo
+      fonts:
+        - asset: assets/fonts/Heebo/Heebo-Regular.ttf
